+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 **Battlegrounds**:
 - Fixed Tier7 hero picking stats not updating when rerolling a skin.
+- Added counters for Beetles stats, Spells played and Undead attack bonus.
 
 ## **Release v1.37.5 - 2024-12-03**
 **Updated for Hearthstone 31.2.0**